fix: move functions and files
