+++ src/i18n/locales/en.json
@@ -21,10 +21,12 @@
       "name": "Name",
       "incomingFlights": "Incoming Flights",
       "outgoingFlights": "Outgoing Flights",
+      "actions": {
       "areYouAbsolutelySure": "Are you absolutely sure?",
       "thisActionCantBeUndone": "This action cannot be undone. Are you sure you want to permanently delete { name }?",
       "deletionSuccess": "{ name } was deleted successfully!",
       "deletionError": "There was an error deleting { name }."
+      }
     }
   },
   "validation": {
