+++ src/services/index.ts
@@ -1,3 +1,5 @@
+export * from "./airlines";
 export * from "./auth";
 export * from "./cities";
+//export * from "./flights";
 export * from "./users";
